02. Web App/requirements_app.txt
fastapi>=0.115.0
uvicorn>=0.31.0
tensorflow-cpu>=2.18.0
opencv-python-headless>=4.10.0.84
Pillow>=11.0.0
python-multipart>=0.0.12